refactor: Email DTO, Service 패키지 변경
