+++ src/main/java/com/example/demo/dto/email/EmailSendRequest.java
@@ -1,4 +1,4 @@
-package com.example.demo.dto;
+package com.example.demo.dto.email;
 
 import java.util.List;
