--- src/main/java/com/example/demo/service/MailService.java
@@ -1,9 +1,0 @@
-package com.example.demo.service;
-
-import com.example.demo.dto.EmailSendRequest;
-import com.example.demo.dto.EmailTemplateResponse;
-
-public interface MailService {
-
-	EmailTemplateResponse sendEmail(EmailSendRequest request);
-}
