+++ src/main/java/com/example/demo/service/email/MailService.java
@@ -1,0 +1,9 @@
+package com.example.demo.service.email;
+
+import com.example.demo.dto.email.EmailSendRequest;
+import com.example.demo.dto.email.EmailTemplateResponse;
+
+public interface MailService {
+
+	EmailTemplateResponse sendEmail(EmailSendRequest request);
+}
