+++ src/main/java/com/example/demo/service/email/MailServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.demo.service;
+package com.example.demo.service.email;
 
 import static com.example.demo.exception.ExceptionType.*;
 
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Service;
 import com.example.demo.domain.MemberImpl;
 import com.example.demo.domain.email.EmailSender;
 import com.example.demo.domain.email.EmailTemplate;
-import com.example.demo.dto.EmailSendRequest;
-import com.example.demo.dto.EmailTemplateResponse;
+import com.example.demo.dto.email.EmailSendRequest;
+import com.example.demo.dto.email.EmailTemplateResponse;
 import com.example.demo.exception.CafegoryException;
 import com.example.demo.repository.MemberRepository;
